Rewrite JSON interface.
  - We now use `json_data_lookup=scope=value`
  - This means we can use commas in our filters...
  - We now include the lookup in the key.
  - If the last element of the key is not a lookup, we add __exact.

NOTE: If the user wants to search in a key that shares a name with a lookup type, they are... Not going to have a good time. Ie, you have the following structure: `{ "key" { "exact": "value" } }`, filtering on `key__exact=value` will fail, as the value of `key` is a list (`{"exact": "value"}`).

To get around this, one has to use the `key` functionality in a filter akin to this: `filter(weird__key='exact', weird__exact='value')`. This is however not supported. See #76.

(Also, add debug output with request data if a test fails.)