fix(testing): Ensure testing leaves nothing behind (#157)
Under GitHub Actions we sometimes get attachment deletion failures
in sqlite environment.

Removing only the attachments and not the folders is a bit safer in
general, and removing the testing of the tearDownModule removes
race conditions when running with high parallelization, which is
probably why this was only seen during sqlite testing.

Closes #156